services/authentication/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.25.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==2.5.3
PyJWT==2.8.0
python-multipart==0.0.6
requests==2.31.0
